# Conflicts:
#	src/args.cpp
#	src/args.h
#	src/options.cpp
#	src/parse_frame.cpp
#	src/prototypes.h
#	src/tokenize_cleanup.cpp
#	src/uncrustify.cpp
#	src/unicode.cpp
#	src/width.cpp

+++ include/uncrustify_types.h
@@ -327,7 +327,7 @@ struct chunk_tag_t
 {
    const char *tag;
    c_token_t  type;
-   int        lang_flags;
+   size_t     lang_flags;
 };
 
 
@@ -395,7 +395,7 @@ struct cp_data_t
    file_mem_t     oc_msg_hdr;     // for cmt_insert_oc_msg_header
    file_mem_t     class_hdr;      // for cmt_insert_class_header
 
-   int            lang_flags;     // LANG_xxx
+   size_t         lang_flags;     // LANG_xxx
    bool           lang_forced;
 
    bool           unc_off;
